+++ amazon-chime-sdk/src/main/java/com/amazonaws/services/chime/sdk/meetings/audiovideo/video/capture/DefaultCameraCaptureSource.kt
@@ -195,7 +195,7 @@ class DefaultCameraCaptureSource @JvmOverloads constructor(
             handleCameraCaptureFail(CaptureSourceError.PermissionError)
             throw SecurityException("Missing necessary camera permissions")
         }
-
+        stop()
         logger.info(TAG, "Camera capture start requested with device: $device")
         val device = device ?: run {
             logger.info(TAG, "Cannot start camera capture with null device")