+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ### Added
 * Added `audioStreamType` in `AudioVideoConfiguration` for supporting audio stream configuration.
 
+### Fixed
+* Fixed calling start multiple times without stop crashes video in some phones. (Issue #356)
+
 ## [0.15.4] - 2022-04-21
 
 ## [0.15.3] - 2022-04-07
@@ -33,9 +36,6 @@
 ### Added
 * Added the meetingStartDurationMs event in ingestionEvents to record the time that elapsed between the start request and the beginning of the meeting.
 * Added priority based downlink policy to control the way how a recipient subscribes to the remote video sources
-
-### Fixed
-* Fixed calling start multiple times without stop crashes video in some phones. (Issue #356)
 
 ## [0.14.3] - 2022-02-10