+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ### Added
 * Added `audioStreamType` in `AudioVideoConfiguration` for supporting audio stream configuration.
 
+### Fixed
+* Fixed calling start multiple times without stop crashes video in some phones. (Issue #356)
+
 ## [0.15.4] - 2022-04-21
 
 ## [0.15.3] - 2022-04-07